+++ src/components/POLine/POLineForm.js
@@ -1,6 +1,6 @@
 import React, { useCallback, useEffect } from 'react';
 import PropTypes from 'prop-types';
-import { get, mapValues } from 'lodash';
+import { get, mapValues, pick } from 'lodash';
 import { FormattedMessage } from 'react-intl';
 import { useHistory } from 'react-router';
 
@@ -58,7 +58,7 @@ import getOrderTemplateValue from '../Utils/getOrderTemplateValue';
 import calculateEstimatedPrice from './calculateEstimatedPrice';
 import styles from './POLineForm.css';
 
-const GAME_CHANGER_FIELDS = ['isPackage', 'orderFormat', 'checkinItems', 'packagePoLineId'];
+const GAME_CHANGER_FIELDS = ['isPackage', 'orderFormat', 'checkinItems', 'packagePoLineId', 'instanceId'];
 
 function POLineForm({
   form: { change, batch },
@@ -87,6 +87,21 @@ function POLineForm({
   });
   const lineId = get(initialValues, 'id');
   const saveBtnLabelId = isCreateAnotherChecked ? 'save' : 'saveAndClose';
+  const initialInventoryData = (
+    !lineId && templateValue.id
+      ? {
+        ...pick(templateValue, [
+          'instanceId',
+          'titleOrPackage',
+          'publisher',
+          'publicationDate',
+          'edition',
+          'contributors',
+          'details.productIds',
+        ]),
+      }
+      : {}
+  );
 
   useEffect(() => {
     setTimeout(() => {
@@ -341,7 +356,7 @@ function POLineForm({
                         change={change}
                         batch={batch}
                         identifierTypes={identifierTypes}
-                        initialValues={initialValues}
+                        initialValues={{ ...initialValues, ...initialInventoryData }}
                         stripes={stripes}
                       />
                     </Accordion>